4-Data_Engineering/3-Streamlit/Taller_Streamlit/streamlit/requirements.txt
pandas==2.3.3
matplotlib==3.10.0
pillow==12.0.0